requirements.txt
streamlit
scikit-learn
pandas
joblib
numpy
google-oauth2-tool
google-auth
google-auth-oauthlib
google-api-python-client
pyinstaller
